+++ src/main/java/Simulation/Assets.java
@@ -1,20 +1,20 @@
 package Simulation;
 
 public class Assets {
-    public final int cash;
-    public final int stocks;
+    public final long cash;
+    public final long stocks;
 
-    public Assets(int cash, int stocks) {
+    public Assets(long cash, long stocks) {
         if(cash < 0 || stocks < 0) { throw new RuntimeException("Can't create assets with negative values."); }
         this.cash = cash;
         this.stocks = stocks;
     }
 
-    public Assets add(int cash, int stocks) {
+    public Assets add(long cash, long stocks) {
         return new Assets(this.cash + cash, this.stocks + stocks);
     }
 
-    public int toCash(int stocksValue) {
+    public long toCash(long stocksValue) {
         return this.cash + this.stocks * stocksValue;
     }
