+++ src/main/java/Simulation/Simulation.java
@@ -13,20 +13,20 @@ import java.util.stream.Collectors;
 
 public class Simulation {
 
-    private Map<Observable, List<Pair<Integer, Double>>> observables;
+    private Map<Observable, List<Pair<Long, Double>>> observables;
     private List<Agent> agents;
     private OrdersBook ordersBook;
-    private Integer turn;
+    private int turn;
 
 
     public final Configuration configuration;
 
-    private final Map<Observable, Consumer<List<Pair<Integer, Double>>>> observablesUpdateFunctions;
+    private final Map<Observable, Consumer<List<Pair<Long, Double>>>> observablesUpdateFunctions;
 
     public Simulation(Configuration configuration)
     {
         this.configuration = configuration;
-        Map<Observable, Consumer<List<Pair<Integer, Double>>>> tmpFunctionMap = new HashMap<>();
+        Map<Observable, Consumer<List<Pair<Long, Double>>>> tmpFunctionMap = new HashMap<>();
         tmpFunctionMap.put(Observable.ASK_PRICE_HISTORY, this::updateAskPriceHistory);
         tmpFunctionMap.put(Observable.BID_PRICE_HISTORY, this::updateBidPriceHistory);
         tmpFunctionMap.put(Observable.MARKET_DEPTH, this::updateMarketDepth);
@@ -36,15 +36,15 @@ public class Simulation {
         initialize();
     }
 
-    public Map<Observable, List<Pair<Integer, Double>>> getObservables() { return observables; }
-    public List<Pair<Integer, Double>> getObservable(Observable ob) { return observables.get(ob); }
+    public Map<Observable, List<Pair<Long, Double>>> getObservables() { return observables; }
+    public List<Pair<Long, Double>> getObservable(Observable ob) { return observables.get(ob); }
     public OrdersBook getOrdersBook() { return ordersBook; }
 
     public void nextTurn()
     {
         Collections.shuffle(agents);
         ordersBook.clearTransactions();
-        int agent_count = 0;
+        Long agent_count = 0L;
         for(Agent agent : agents)
         {
             agent_count++;
@@ -79,10 +79,10 @@ public class Simulation {
 
         for(int i = 0; i < this.configuration.NUMBER_OF_AGENTS; i++) {
             this.agents.add(new Agent.Builder(this)
-                .perception(configuration.PERCEPTION)
-                .tactic(tactic)
-                .assets(new Assets(configuration.INITIAL_CASH, configuration.INITIAL_STOCKS))
-                .build());
+                    .perception(configuration.PERCEPTION)
+                    .tactic(tactic)
+                    .assets(new Assets(configuration.INITIAL_CASH, configuration.INITIAL_STOCKS))
+                    .build());
         }
     }
 
@@ -94,9 +94,9 @@ public class Simulation {
         }
     }
 
-    private void updateBidPriceHistory(List<Pair<Integer, Double>> s)
+    private void updateBidPriceHistory(List<Pair<Long, Double>> s)
     {
-        int lastTime = s.size() == 0 ? 0 : s.get(s.size() - 1).getKey();
+        Long lastTime = s.size() == 0 ? 0L : s.get(s.size() - 1).getKey();
         if (ordersBook.getBuyOrders().size() == 0) {
             s.add(new Pair<>(lastTime + 1, 0.0));
         } else {
@@ -104,9 +104,9 @@ public class Simulation {
         }
     }
 
-    private void updateAskPriceHistory(List<Pair<Integer, Double>> s)
+    private void updateAskPriceHistory(List<Pair<Long, Double>> s)
     {
-        int lastTime = s.size() == 0 ? 0 : s.get(s.size() - 1).getKey();
+        long lastTime = s.size() == 0L ? 0L : s.get(s.size() - 1).getKey();
         if (ordersBook.getSellOrders().size() == 0) {
             s.add(new Pair<>(lastTime + 1, 0.0));
         } else {
@@ -114,22 +114,23 @@ public class Simulation {
         }
     }
 
-    private void updateMarketDepth(List<Pair<Integer, Double>> s)
+    private void updateMarketDepth(List<Pair<Long, Double>> s)
     {
         s.clear();
         List<BuyOffer> bo = ordersBook.getBuyOrders().stream()
-            .sorted(Comparator.comparing((BuyOffer x) -> x.price).reversed())
-            .collect(Collectors.toList());
+                .sorted(Comparator.comparing((BuyOffer x) -> x.price).reversed())
+                .collect(Collectors.toList());
         List<SellOffer> so = ordersBook.getSellOrders().stream()
-            .sorted(Comparator.comparing((SellOffer x) -> x.price))
-            .collect(Collectors.toList());
-        int cumBuy = 0, lastBuyPrice = -1;
+                .sorted(Comparator.comparing((SellOffer x) -> x.price))
+                .collect(Collectors.toList());
+        Long cumBuy = 0L, lastBuyPrice = -1L;
         for(BuyOffer offer : bo)
         {
             cumBuy += offer.getStockQuantity();
             if (offer.price == lastBuyPrice)
             {
                 s.remove(s.size() - 1);
+
                 s.add(new Pair<>(offer.price, (double)cumBuy));
             }
             else
@@ -138,7 +139,7 @@ public class Simulation {
             }
             lastBuyPrice = offer.price;
         }
-        int cumSell = 0, lastSellPrice = -1;
+        Long cumSell = 0L, lastSellPrice = -1L;
         for(SellOffer offer : so)
         {
             cumSell += offer.getStockQuantity();
@@ -155,44 +156,44 @@ public class Simulation {
         }
     }
 
-    private void updateLogReturns(List<Pair<Integer, Double>> s)
+    private void updateLogReturns(List<Pair<Long, Double>> s)
     {
-        int lastTime = s.size() == 0 ? 0 : s.get(s.size() - 1).getKey();
-        List<Pair<Integer, Double>> askHistory = observables.get(Observable.ASK_PRICE_HISTORY);
-        List<Pair<Integer, Double>> bidHistory = observables.get(Observable.BID_PRICE_HISTORY);
+        Long lastTime = s.size() == 0 ? 0L : s.get(s.size() - 1).getKey();
+        List<Pair<Long, Double>> askHistory = observables.get(Observable.ASK_PRICE_HISTORY);
+        List<Pair<Long, Double>> bidHistory = observables.get(Observable.BID_PRICE_HISTORY);
         double lastAskPrice = askHistory.size() > 0
-            ? askHistory.get(askHistory.size() - 1).getValue()
-            : configuration.INITIAL_PRICE;
+                ? askHistory.get(askHistory.size() - 1).getValue()
+                : configuration.INITIAL_PRICE;
         double lastBidPrice = bidHistory.size() > 0 ?
-            bidHistory.get(bidHistory.size() - 1).getValue()
-            : configuration.INITIAL_PRICE;
+                bidHistory.get(bidHistory.size() - 1).getValue()
+                : configuration.INITIAL_PRICE;
         double secondLastAskPrice = askHistory.size() > 1
-            ? askHistory.get(askHistory.size() - 2).getValue()
-            : configuration.INITIAL_PRICE;
+                ? askHistory.get(askHistory.size() - 2).getValue()
+                : configuration.INITIAL_PRICE;
         double secondLastBidPrice = bidHistory.size() > 1
-            ? bidHistory.get(bidHistory.size() - 2).getValue()
-            : configuration.INITIAL_PRICE;
+                ? bidHistory.get(bidHistory.size() - 2).getValue()
+                : configuration.INITIAL_PRICE;
         double secondLastMidPrice = 0.5 * (Math.log(secondLastAskPrice) + Math.log(secondLastBidPrice));
         double lastMidPrice = 0.5 * (Math.log(lastAskPrice) + Math.log(lastBidPrice));
         s.add(new Pair<>(lastTime + 1, Math.abs( lastMidPrice - secondLastMidPrice)));
     }
 
-    private void updateLogSpread(List<Pair<Integer, Double>> s)
+    private void updateLogSpread(List<Pair<Long, Double>> s)
     {
-        int lastTime = s.size() == 0 ? 0 : s.get(s.size() - 1).getKey();
-        List<Pair<Integer, Double>> askHistory = observables.get(Observable.ASK_PRICE_HISTORY);
-        List<Pair<Integer, Double>> bidHistory = observables.get(Observable.BID_PRICE_HISTORY);
+        Long lastTime = s.size() == 0 ? 0L : s.get(s.size() - 1).getKey();
+        List<Pair<Long, Double>> askHistory = observables.get(Observable.ASK_PRICE_HISTORY);
+        List<Pair<Long, Double>> bidHistory = observables.get(Observable.BID_PRICE_HISTORY);
         double lastAskPrice = askHistory.size() > 0
-            ? askHistory.get(askHistory.size() - 1).getValue()
-            : configuration.INITIAL_PRICE;
+                ? askHistory.get(askHistory.size() - 1).getValue()
+                : configuration.INITIAL_PRICE;
         double lastBidPrice = bidHistory.size() > 0 ?
-            bidHistory.get(bidHistory.size() - 1).getValue()
-            : configuration.INITIAL_PRICE;
+                bidHistory.get(bidHistory.size() - 1).getValue()
+                : configuration.INITIAL_PRICE;
         double logSpread = Math.log(lastAskPrice) - Math.log(lastBidPrice);
         s.add(new Pair<>(lastTime + 1, logSpread));
     }
 
-    public Integer getTurn() {
+    public int getTurn() {
         return turn;
     }
 }
